rename: 누락된 파일명 케밥 형식으로 변경
